Pull some functionality into support files
