+++ vesuvius/src/vesuvius/utils/transforms/find_transform.py
@@ -1,26 +1,15 @@
 """Find the transform between two provided volumes."""
 
 import argparse
-import json
 from pathlib import Path
 import sys
-from typing import Optional, NamedTuple
-from urllib.parse import urljoin, urlparse
 import webbrowser
 
 import neuroglancer
 import numpy as np
-import requests
-import zarr
 
-
-class Dimensions(NamedTuple):
-    """Structure for volume dimensions with x, y, z coordinates and voxel size."""
-
-    voxels_x: int
-    voxels_y: int
-    voxels_z: int
-    voxel_size_um: float
+from registration import align_zarrs
+from transform_utils import Dimensions, get_volume_dimensions
 
 
 GREEN_SHADER = """
@@ -34,97 +23,6 @@ void main() {
     emitRGB(vec3(toNormalized(getDataValue()), 0, toNormalized(getDataValue())));
 }
 """
-
-
-def sanity_check_zarr_store(store):
-    """Check if the store is a valid OME-ZARR file."""
-    try:
-        # Basic functionality checks
-        assert store is not None
-        assert hasattr(store, "keys")
-        assert hasattr(store, "attrs")
-    except AssertionError as e:
-        raise ValueError(f"Invalid OME-ZARR file: {store.path}") from e
-
-
-def get_volume_dimensions(
-    path: str, provided_voxel_size: Optional[float] = None
-) -> Dimensions:
-    """Get volume dimensions (from Zarr array shape).
-
-    Also get voxel size in microns (from metadata.json if it exists, otherwise use provided value).
-    If both are provided, make sure they are the same.
-    """
-
-    # Get volume dimensions in voxels
-    with zarr.open(path, mode="r") as store:
-        sanity_check_zarr_store(store)
-        voxels_x, voxels_y, voxels_z = store["0"].shape
-
-    # Get voxel size from metadata.json if it exists
-    # Check if path is a URL or local path
-    parsed = urlparse(path)
-    is_remote = parsed.scheme in ("http", "https")
-
-    metadata = None
-    metadata_path = None
-
-    if is_remote:
-        # Remote path
-        metadata_url = urljoin(path + "/", "metadata.json")
-        try:
-            response = requests.get(metadata_url, timeout=10)
-            if response.status_code == 200:
-                metadata = response.json()
-        except (
-            requests.RequestException,
-            json.JSONDecodeError,
-        ) as e:
-            raise RuntimeError(
-                f"Could not fetch/parse metadata from {metadata_url}: {e}"
-            )
-        metadata_path = metadata_url
-    else:
-        # Local path
-        metadata_path = Path(path) / "metadata.json"
-        if metadata_path.exists():
-            try:
-                with metadata_path.open("r") as f:
-                    metadata = json.load(f)
-            except json.JSONDecodeError as e:
-                raise RuntimeError(
-                    f"Could not parse metadata from {metadata_path}: {e}"
-                )
-
-    if metadata is not None:
-        metadata_voxel_size_mm = (
-            metadata.get("scan")
-            .get("tomo")
-            .get("acquisition")
-            .get("detector")
-            .get("samplePixelSize")
-        )
-        metadata_voxel_size_um = metadata_voxel_size_mm * 1000
-        if provided_voxel_size is not None:
-            assert (
-                metadata_voxel_size_um == provided_voxel_size
-            ), "Voxel size from metadata.json and provided voxel size do not match"
-        return Dimensions(
-            voxels_x=voxels_x,
-            voxels_y=voxels_y,
-            voxels_z=voxels_z,
-            voxel_size_um=metadata_voxel_size_um,
-        )
-    else:
-        assert (
-            provided_voxel_size is not None
-        ), f"No metadata.json found at {metadata_path} and no voxel size provided directly"
-        return Dimensions(
-            voxels_x=voxels_x,
-            voxels_y=voxels_y,
-            voxels_z=voxels_z,
-            voxel_size_um=provided_voxel_size,
-        )
 
 
 def add_moving_and_fixed_layers(
@@ -370,6 +268,24 @@ def read_transform_from_file(input_path: str) -> np.ndarray:
         return np.loadtxt(f, dtype=np.float64)
 
 
+def fine_align(_):
+    """Run zarr alignment and update the moving layer with the result."""
+    with viewer.txn() as state:
+        current_transform = state.layers["moving"].layer.source[0].transform.matrix
+        # Add homogeneous coordinate
+        current_transform = np.concatenate([current_transform, [[0, 0, 0, 1]]], axis=0)
+
+        # Run alignment
+        print("Running zarr alignment...")
+        refined_transform = align_zarrs(args.fixed, args.moving, current_transform)
+
+        # Apply the refined transform
+        # Remove homogeneous coordinate
+        refined_transform = refined_transform[:-1, :]
+        state.layers["moving"].layer.source[0].transform.matrix = refined_transform
+        print("Alignment complete - transform updated")
+
+
 def write_current_transform(_):
     """Write the current transform and print the shareable URL."""
     with viewer.txn() as state:
@@ -398,6 +314,7 @@ def add_actions_and_keybinds(viewer: neuroglancer.Viewer) -> None:
 
     viewer.actions.add("toggle-color", toggle_color)
     viewer.actions.add("write-transform", write_current_transform)
+    viewer.actions.add("fine-align", fine_align)
     viewer.actions.add("rot-x-plus-small", _make_rotator("x", SMALL_ROTATE_STEP))
     viewer.actions.add("rot-x-minus-small", _make_rotator("x", -SMALL_ROTATE_STEP))
     viewer.actions.add("rot-y-plus-small", _make_rotator("y", SMALL_ROTATE_STEP))
@@ -423,6 +340,7 @@ def add_actions_and_keybinds(viewer: neuroglancer.Viewer) -> None:
     with viewer.config_state.txn() as s:
         s.input_event_bindings.viewer["keyc"] = "toggle-color"
         s.input_event_bindings.viewer["keyw"] = "write-transform"
+        s.input_event_bindings.viewer["keyf"] = "fine-align"
         s.input_event_bindings.viewer["alt+keya"] = "rot-x-plus-small"
         s.input_event_bindings.viewer["alt+keyq"] = "rot-x-minus-small"
         s.input_event_bindings.viewer["alt+keys"] = "rot-y-plus-small"
@@ -511,6 +429,8 @@ if __name__ == "__main__":
 
     webbrowser.open_new(viewer.get_viewer_url())
 
-    # Allow clicking to set points
-    # Have some mechanism to have current active layer
-    # Once enough points: find affine transform
+    # TODO
+    # Create registration file with stub registration method and map it in this file to a keybind
+    # Figure out what resolutions are available
+    # If low enough, try SimpleITK registration using lowest resolution
+    # Then try with higher and/or chunking?
