+++ vesuvius/src/vesuvius/utils/transforms/registration.py
@@ -1,0 +1,19 @@
+import numpy as np
+from typing import Tuple, Optional
+
+
+def align_zarrs(zarr1_path: str, zarr2_path: str, initial_transform: np.ndarray) -> np.ndarray:
+    """
+    Align two zarr datasets given their paths and an initial transform.
+    
+    Args:
+        zarr1_path: Path to the first zarr dataset
+        zarr2_path: Path to the second zarr dataset  
+        initial_transform: Initial affine transformation matrix
+        
+    Returns:
+        Refined affine transformation matrix
+    """
+    # TODO: Implement zarr alignment algorithm
+    # For now, return the initial transform unchanged
+    return initial_transform
